:recycle: Game play 로직 리팩토링
